% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePQI.R
\name{computePQI}
\alias{computePQI}
\title{Add together two numbers}
\usage{
computePQI(data = data, pqi = c(1, 3, 5), append = FALSE,
  diagcode = "dx", proccode = "px", agevar = "age", version = "5.0")
}
\arguments{
\item{data}{the inpute data with diagnosis and procedure codes}

\item{pqi}{Which PQI's to calculate}

\item{append}{Should we append the output to the original data file.  Default is FALSE.}

\item{diagcode}{test}

\item{proccode}{test}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)


}

